services/users/requirements.txt
Flask==1.1.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
gunicorn==19.9.0
coverage==4.5.4
flake8==3.7.8
Flask-DebugToolbar==0.10.1
Flask-Cors==3.0.8 
Flask-Migrate==2.5.2
Flask-Bcrypt==0.7.1
PyJWT==1.7.1
Werkzeug==1.0.0
psycopg2-binary==2.8.4
